+++ RP_Notify/Resources/config.ini
@@ -3,7 +3,6 @@ ShowOnNewSong=true
 LargeAlbumArt=true
 ShowSongRating=true
 [AppSettings]
-LeaveShorcutInStartMenu=true
 EnableFoobar2000Watcher=false
 EnableMusicBeeWatcher=false
 EnableRpOfficialTracking=false
